+++ src/moduleStore/jsCache.js
@@ -22,7 +22,7 @@ export const getJSForModule = async (moduleName) => {
   if (cache[moduleName] && moduleInfo.version === cache[moduleName].version) {
     return cache[moduleName].js;
   } else {
-    const js = await goosemodScope.cspBypasser.text(moduleInfo.codeURL, false);
+    const js = await goosemodScope.cspBypasser.text(`${moduleInfo.codeURL}?_=${Date.now()}`, false);
 
     goosemodScope.moduleStoreAPI.jsCache.updateCache(moduleName, moduleInfo.version, js);