+++ src/util/discord/cspBypasser.js
@@ -1,113 +1,74 @@
 let goosemodScope = {};
 
+let noCSPFetch;
+
 export default {
   setThisScope: (scope) => {
     goosemodScope = scope;
   },
 
-  frame: document.createElement('object'),
+  noCSPFetch: undefined,
 
   init: async () => {
-    goosemodScope.cspBypasser.frame.data = location.origin;
-    document.body.appendChild(goosemodScope.cspBypasser.frame);
+    if (window.goosemod_noCSPFetch) {
+      noCSPFetch = window.goosemod_noCSPFetch;
 
-    let script = document.createElement('script');
+      delete window.goosemod_noCSPFetch;
+
+      return;
+    }
+
+    let el = document.createElement('object');
+
+    el.data = location.origin;
+    document.body.appendChild(el);
+
+    /*let script = document.createElement('script');
     script.type = 'text/javascript';
 
     let code = `
-    window.addEventListener('message', async (e) => {
-      const {url, type, useCORSProxy} = e.data;
+window.addEventListener('message', async (e) => {
+  const { url, type, useCORSProxy, useCache } = e.data;
 
-      if (!url) return;
+  if (!url) return;
 
-      const proxyURL = useCORSProxy ? \`https://cors-anywhere.herokuapp.com/\${url}\` : url;
+  const req = await fetch(url, {
+    cache: useCache ? 'default' : 'no-store'
+  });
+
+  e.source.postMessage({ verify: url, data: await req[type]() });
+});`;
+
+    script.appendChild(document.createTextNode(code));
+
+    goosemodScope.cspBypasser.frame.contentDocument.head.appendChild(script);*/
+
+    noCSPFetch = el.contentWindow.fetch;
+  },
+
+
+  fetch: (url, type, useCORSProxy = true) => {
+    return new Promise(async (finalRes) => {
+      const sendURL = useCORSProxy ? `https://cors-anywhere.herokuapp.com/${url}` : url;
+
+      const sendType = type === 'img' ? 'blob' : type;
+
+      const req = await noCSPFetch(sendURL);
+      const resp = await req[sendType]();
 
       if (type === 'img') {
-        let canvas = document.createElement('canvas');
-        let ctx = canvas.getContext('2d');
+        resp = URL.createObjectURL(resp);
+      }
 
-        let img = new Image();
-        img.src = proxyURL;
-        img.crossOrigin = 'anonymous';
-
-        img.onload = () => {
-          canvas.width = img.width;
-          canvas.height = img.height;
-
-          ctx.drawImage(img, 0, 0);
-
-          e.source.postMessage({ verify: url, data: canvas.toDataURL("image/png") });
-        };
-
-        return;
-      }       
-      
-      const req = await fetch(proxyURL, {
-        cache: 'no-store'
-      });
-
-      e.source.postMessage({ verify: url, data: await req[type]() });
-    });`;
-
-    script.appendChild(document.createTextNode(code));
-
-    goosemodScope.cspBypasser.frame.contentDocument.head.appendChild(script);
-  },
-
-  runCode: (code) => {
-    let script = document.createElement('script');
-    script.type = 'text/javascript';
-
-    script.appendChild(document.createTextNode(code));
-
-    goosemodScope.cspBypasser.frame.contentDocument.head.appendChild(script);
-  },
-
-  json: (url, useCORSProxy = true) => {
-    return new Promise((res) => {
-      goosemodScope.cspBypasser.frame.contentWindow.postMessage({url, type: 'json', useCORSProxy});
-
-      window.addEventListener('message', async (e) => {
-        if (e.data.verify !== url) return;
-
-        res(e.data.data);
-      });
+      return finalRes(resp);
     });
   },
 
-  text: (url, useCORSProxy = true) => {
-    return new Promise((res) => {
-      goosemodScope.cspBypasser.frame.contentWindow.postMessage({url, type: 'text', useCORSProxy});
+  json: (url, useCORSProxy) => goosemodScope.cspBypasser.fetch(url, 'json', useCORSProxy),
 
-      window.addEventListener('message', async (e) => {
-        if (e.data.verify !== url) return;
+  text: (url, useCORSProxy) => goosemodScope.cspBypasser.fetch(url, 'text', useCORSProxy),
 
-        res(e.data.data);
-      });
-    });
-  },
+  blob: (url, useCORSProxy) => goosemodScope.cspBypasser.fetch(url, 'blob', useCORSProxy),
 
-  blob: (url, useCORSProxy = true) => {
-    return new Promise((res) => {
-      goosemodScope.cspBypasser.frame.contentWindow.postMessage({url, type: 'blob', useCORSProxy});
-
-      window.addEventListener('message', async (e) => {
-        if (e.data.verify !== url) return;
-
-        res(e.data.data);
-      });
-    });
-  },
-
-  image: (url, useCORSProxy = true) => {
-    return new Promise((res) => {
-      goosemodScope.cspBypasser.frame.contentWindow.postMessage({url, type: 'img', useCORSProxy});
-
-      window.addEventListener('message', async (e) => {
-        if (e.data.verify !== url) return;
-
-        res(e.data.data);
-      });
-    });
-  },
+  image: (url, useCORSProxy) => goosemodScope.cspBypasser.fetch(url, 'img', useCORSProxy),
 };